+++ src/index.tsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { wrap } from 'comlink';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -18,13 +17,3 @@ root.render(
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-
-const doLongRunningWorkWithWorker = () => {
-  const worker = new Worker(new URL('./worker/demo.worker.ts', import.meta.url), {
-    type: 'module'
-  })
-  const workerApi = wrap<import('./worker/demo.worker').DemoWorker>(worker);
-  workerApi.doLongRunningWork()
-}
-
-doLongRunningWorkWithWorker()
