+++ src/worker/demo.worker.ts
@@ -1,11 +1,11 @@
 import { expose } from 'comlink'
-import { doLongRunningWork } from '../longRunningWork'
+import { doLongRunningWork } from '../utils/longRunningWork'
 
 const exports = {
   doLongRunningWork
 }
 
 // create the public facing api type
-export type DemoWorker = typeof exports
+export type MyWorker = typeof exports
 
 expose(exports)
