+++ tsconfig.json
@@ -20,7 +20,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": [
-    "src/*"
-  ]
+  "include": ["src/*", "src/utils/longRunningWork.ts"]
 }
